# docs/spec/openapi.yaml
openapi: "3.1.0"
info:
  title: RuneScape API Wrapper
  version: 1.0.0
  description: >
    API for fetching Old School RuneScape player stats via the public hiscores endpoints.

servers:
  - url: https://rs-api.dedeen.dev/api/v1/osrs
    description: Production server

paths:
  /stats:
    get:
      summary: Fetch a player's stats
      description: >
        Retrieves detailed hiscore data for a given player from the selected highscore type.
      operationId: getPlayer
      parameters:
        - name: name
          in: query
          description: The RuneScape username to look up.
          required: true
          schema:
            type: string
            example: "rattesjtaart"
        - name: highscore
          in: query
          description: >
            The highscore category to query. Defaults to NORMAL if not provided.
          required: false
          schema:
            type: string
            enum:
              - NORMAL
              - IRONMAN
              - HARDCORE_IRONMAN
              - ULTIMATE_IRONMAN
              - DEADMAN
              - SEASONAL
              - TOURNAMENT
              - FRESH_START_WORLD
              - PURE
              - LEVEL_3_SKILLER
            default: NORMAL
      responses:
        '200':
          description: Player stats successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '400':
          description: Invalid input (missing or invalid parameters).
        '404':
          description: Player not found.
        '500':
          description: Unexpected error occurred during retrieval.

components:
  schemas:
    Player:
      type: object
      description: Player object containing name, metadata, and stats.
      properties:
        name:
          type: string
          example: "rattesjtaart"
        metadata:
          type: object
          additionalProperties:
            type: string
          example:
            accountType: "NORMAL"
            lastUpdated: "2025-10-13T10:00:00Z"
        stats:
          type: object
          additionalProperties:
            type: string
          example:
            Attack: "99"
            Defence: "85"
            Hitpoints: "99"